Start
Add the jetpack block availability when registering a block.
This way we can display a nice message to the user on how to get the block if they really want to use it.

Add simple payment block support and fix bug

Add check via blocks_index?

Add filters so that we only load the presets when on the jetpack side since we can load the .com blocks using a different filter

Bug fix.

Make it work with the new block manifest file PR.

https://github.com/Automattic/wp-calypso/pull/28250/files

Remove simple payment

Register the 'pulicize' block.

Use the jetpack_constants class

also include old version of state, so we don't introduce breaking changes to bleeding edge